+++ src/apis/accommodation/dto/response/get-accommodation-list.response.dto.ts
