+++ src/apis/accommodation/dto/response/index.ts
@@ -1,5 +1,8 @@
 import GetAccommodationListResponseDto from "./get-accommodation-list.response.dto"
+import ResponseDto from "./response.dto"
+
 
 export type{
-    GetAccommodationListResponseDto
+    GetAccommodationListResponseDto,
+    ResponseDto
 }
