--- src/apis/dto/request/accommodation/index.ts
@@ -1,4 +1,0 @@
-import SearchedAccommodationRequestDto from "./search-accommodation.request.dto";
-import FilterRequestDto from "./filter.request.dto";
-
-export type { SearchedAccommodationRequestDto, FilterRequestDto };
