--- src/apis/dto/request/auth/guest/g-sign-up.request.dto.ts
@@ -1,9 +1,0 @@
-// interface: 회원가입 Reuqest Body Dto //
-
-export default interface GuestSignUpRequestDto {
-    guestId: string;
-    guestPassword: string;
-    guestName: string;
-    telNumber: string;
-    joinPath: string;
-}
