--- src/apis/dto/request/auth/host/h-id-check.requst.dto.ts
@@ -1,4 +1,0 @@
-// interface: 중복 확인 Request Body Dto //
-export default interface HostIdCheckRequestDto {
-    hostId : string;
-}
