--- src/apis/dto/request/auth/host/h-sign-up.request.dto.ts
@@ -1,13 +1,0 @@
-// interface: 회원가입 Reuqest Body Dto //
-
-export default interface HostSignUpRequestDto {
-    hostId: string;
-    hostPassword: string;
-    hostName: string;
-    businessNumber: string;
-    businessImage: string;
-    telNumber: string;
-    businessName: string;
-    businessStartDay: string;
-
-}
