--- src/apis/dto/request/auth/index.ts
@@ -1,15 +1,0 @@
-import TelAuthRequestDto from "./tel-auth.request.dto";
-import GuestIdCheckRequestDto from "./guest/g-id-check.requst.dto";
-import GuestSignUpRequestDto from "./guest/g-sign-up.request.dto";
-import HostIdCheckRequestDto from "./host/h-id-check.requst.dto";
-import HostSignUpRequestDto from "./host/h-sign-up.request.dto";
-import TelAuthCheckRequestDto from "./tel-auth-check.request.dto";
-
-export type {
-  TelAuthRequestDto,
-  TelAuthCheckRequestDto,
-  HostSignUpRequestDto,
-  HostIdCheckRequestDto,
-  GuestSignUpRequestDto,
-  GuestIdCheckRequestDto,
-};
