--- src/apis/dto/request/auth/tel-auth-check.request.dto.ts
@@ -1,5 +1,0 @@
-// interface: 전화번호 인증 확인 Request Body Dto //
-export default interface TelAuthCheckRequestDto {
-    telNumber: string;
-    authNumber: string;
-}
