--- src/apis/dto/request/auth/tel-auth.request.dto.ts
@@ -1,4 +1,0 @@
-// interface: 전화번호 인증 Request Body Dto //
-export default interface TelAuthRequestDto {
-    telNumber: string;
-}
