--- src/apis/dto/response/auth/index.ts
@@ -1,5 +1,0 @@
-import SignInResponseDto from "./sign-in.response.dto";
-
-export type {
-    SignInResponseDto
-}
