--- src/apis/dto/response/auth/login/index.ts
@@ -1,5 +1,0 @@
-import LogInResponseDto from "./log-in.response.dto"
-
-export type {
-    LogInResponseDto
-}
