--- src/apis/dto/response/auth/login/log-in.response.dto.ts
@@ -1,6 +1,0 @@
-import ResponseDto from "../../response.dto";
-
-export default interface LogInResponseDto extends ResponseDto {
-    accessToken: string;
-    expiration: number;
-}
