--- src/apis/dto/response/auth/sign-in.response.dto.ts
@@ -1,11 +1,0 @@
-import ResponseDto from "../response.dto";
-
-// interface: 로그인 Response Body Dto //
-export default interface SignInResponseDto extends ResponseDto {
-    accessToken: string;
-    expiration: number;
-}
-
-export default function tm (){
-    
-}
