--- src/apis/dto/response/index.ts
@@ -1,5 +1,0 @@
-import ResponseDto from "./response.dto";
-
-export type {
-    ResponseDto
-}
